#692 moved integration tests
