+++ tests/integration/test_models/test_full_battery_models/test_lithium_ion/test_external/test_external_current_collector.py
@@ -1,0 +1,15 @@
+#
+# Tests for external submodels
+#
+import pybamm
+import unittest
+import numpy as np
+
+
+if __name__ == "__main__":
+    print("Add -v for more debug output")
+    import sys
+
+    if "-v" in sys.argv:
+        debug = True
+    unittest.main()
